backend/app/auth/jwt_utils.py
from datetime import datetime, timedelta
from jose import JWTError, jwt
from ..config.jwt_config import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_MINUTES


def create_access_token(subject: str) -> str:
    """Create a JWT with 'sub' claim (username)."""
    if not SECRET_KEY:
        raise ValueError("JWT_SECRET_KEY is not set")
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode = {"sub": subject, "exp": expire}
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)


def decode_token(token: str) -> dict | None:
    """Decode and verify JWT; return payload or None on failure."""
    if not SECRET_KEY:
        return None
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        return None
