backend/requirements.txt
fastapi==0.128.5
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-dotenv==1.2.1
openai==2.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

